implement as Mul
